% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{get_pos}
\alias{get_pos}
\alias{mut_ref}
\alias{rev_encode}
\title{Auxillary functions}
\usage{
get_pos(chrom)

mut_ref(base)

rev_encode(x)
}
\arguments{
\item{chrom}{A character vector of chromsomes formatted as e.g. "chr1"}

\item{base}{an single DNA nucleotide (ATCG) to mutate}

\item{x}{a vector of numbers from 0 - 2 to encode. ANy other numbers will become no calls ("./.")}
}
\description{
get_pos: get a random postion on a chromosome

mut_ref: mutate a reference base semi randomly, with a Ti/Tv of roughly 2, needs to wrapped in a map/lapply for working with vectors

rev_encode: reverse the encoding of genotypes from numbers to that use in VCF i.e. 0/0, 0/1, 1/1
}
