% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_vcf.R
\name{simulate_vcf}
\alias{simulate_vcf}
\title{Simulate Variants in Hardy Weinberg for the human genome (hg38)}
\usage{
simulate_vcf(num_vars = 500L, num_ppl = 5000L, path = "sim.vcf.gz")
}
\arguments{
\item{num_vars}{The number of variants to simulate, default: 500}

\item{num_ppl}{The number of individuals to simulate, default: 5000}

\item{path}{The output file path, default: sim.vcf.gz}
}
\description{
Simulate Variants in Hardy Weinberg for the human genome (hg38)
}
